+++ packages/core/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @clarigen/core
 
+## 1.0.16-alpha.6
+
+### Patch Changes
+
+- [#37](https://github.com/mechanismHQ/clarigen/pull/37) [`f63a739`](https://github.com/mechanismHQ/clarigen/commit/f63a73930d80bddb83f1ec18e886a5c738e3a8ef) Thanks [@hstove](https://github.com/hstove)! - Fixed a bug in `cvToValue` when changing tuples with responses in values
+
+## 1.0.16-alpha.5
+
+### Patch Changes
+
+- [#37](https://github.com/mechanismHQ/clarigen/pull/37) [`ac1ef2f`](https://github.com/mechanismHQ/clarigen/commit/ac1ef2ffa39dcb580693694b6c98d15194f07910) Thanks [@hstove](https://github.com/hstove)! - Fixes `toCamelCase` to handle UPPER_CASE format
+
+## 1.0.16-alpha.4
+
+### Patch Changes
+
+- [`b969418`](https://github.com/mechanismHQ/clarigen/commit/b9694186e7b9a07368d75b46a66dcfdb1f9357d7) Thanks [@hstove](https://github.com/hstove)! - Improves types in `test` package
+
+## 1.0.16-alpha.3
+
+### Patch Changes
+
+- [`911cc06`](https://github.com/mechanismHQ/clarigen/commit/911cc06c8f293d53c29e2cd5e16eff38a7762053) Thanks [@hstove](https://github.com/hstove)! - Include missing built files
+
+## 1.0.16-alpha.2
+
+### Patch Changes
+
+- [`9a863ad`](https://github.com/mechanismHQ/clarigen/commit/9a863adcb5c94582e369181f2a8773078416f4f4) Thanks [@hstove](https://github.com/hstove)! - Export utils from `test`
+
+## 1.0.16-alpha.1
+
+### Patch Changes
+
+- [`1e52fd6`](https://github.com/mechanismHQ/clarigen/commit/1e52fd6b8278feec80961dcdd1f34ddf393a132f) Thanks [@hstove](https://github.com/hstove)! - Adds raw clarity parsing capabilities to `core` and more helper functions to `test`
+
+## 1.0.16-alpha.0
+
 ## 1.0.15
 
 ### Patch Changes